+++ walkability/core/input.py
@@ -1,13 +1,12 @@
 from enum import Enum
 from typing import Set
 
-from climatoology.utility.Naturalness import NaturalnessIndex
 from pydantic import BaseModel, Field
 
 
 class WalkabilityIndicators(Enum):
     SLOPE = 'Slope'
-    NATURALNESS = 'Naturalness'
+    NATURALNESS = 'Greenness'
     DETOURS = 'Detour Factor'
 
 
@@ -21,18 +20,12 @@ WALKING_SPEED_MAP = {WalkingSpeed.SLOW: 2, WalkingSpeed.MEDIUM: 4, WalkingSpeed.
 
 
 class ComputeInputWalkability(BaseModel):
-    indicators_to_compute: Set[WalkabilityIndicators] = Field(
+    optional_indicators: Set[WalkabilityIndicators] = Field(
         title='Optional indicators',
         description='Computing these indicators for large areas may exceed '
         'the time limit for individual assessments in the Climate Action Navigator.',
         examples=[],
         default=[],
-    )
-    naturalness_index: NaturalnessIndex = Field(
-        title='What to include in naturalness calculation?',
-        description='Choose NDVI to include only vegetation greenness, WATER to include only water bodies, and NATURALNESS to include both.',
-        examples=[NaturalnessIndex.NDVI],
-        default=NaturalnessIndex.NDVI,
     )
 
     @property