+++ CHANGELOG.md
@@ -13,8 +13,11 @@ and this project mostly adheres to [Semantic Versioning](https://semver.org/spec
 
 ### Changed
 
-- Split Not Walkable category into an inaccessible category 'No access' and a category for shared with cars above 50 km/h ([#204](https://gitlab.heigit.org/climate-action/plugins/walkability/-/issues/204))
+- "Not Walkable" paths are now split into two categories: "No access" and "Shared with cars above 50 km/h" ([#204](https://gitlab.heigit.org/climate-action/plugins/walkability/-/issues/204))
 - Updated legend labels for path categories shared with cars to include speed limit instead of category label
+- Renamed input field to select optional indicators, making computation parameters report more understandable ([#224](https://gitlab.heigit.org/climate-action/plugins/walkability/-/issues/224))
+- The Naturalness indicator now includes only NDVI and was thus renamed "Greenness".
+
 - Docker builds now take the commit hash as an argument to avoid version collisions on staging ([#226](https://gitlab.heigit.org/climate-action/plugins/walkability/-/issues/226))
 - CI pipeline now includes test-coverage checks ([#225](https://gitlab.heigit.org/climate-action/plugins/walkability/-/issues/225))
 - Updated ruff and pre-commit configs for more checks ([#181](https://gitlab.heigit.org/climate-action/plugins/walkability/-/issues/181))